Creating Auth Check resource
